Silence warnings from previous commits.
The ioctl ones are *probably* harmless, and silencing them would require
something like Android does, which requires a compiler extension
unavailable in GCC :/.

c.f., https://github.com/rust-lang/libc/issues/1036#issuecomment-565444977

+++ fbink_button_scan.h
@@ -81,8 +81,8 @@
 			({                                                                                               \
 				struct timeval now;                                                                      \
 				gettimeofday(&now, NULL);                                                                \
-				ev.input_event_sec  = now.tv_sec;                                                        \
-				ev.input_event_usec = now.tv_usec;                                                       \
+				ev.input_event_sec  = (unsigned long int) now.tv_sec;                                    \
+				ev.input_event_usec = (unsigned long int) now.tv_usec;                                   \
 				ev.type             = (t);                                                               \
 				ev.code             = (c);                                                               \
 				ev.value            = (v);                                                               \
